write logNumericFacetBreadcrumb explicitly
https://coveord.atlassian.net/browse/KIT-661
