+++ packages/headless/src/features/facets/range-facets/generic/range-facet-analytics-actions.ts
@@ -1,14 +1,6 @@
 import {FacetRangeMetadata} from 'coveo.analytics/src/searchPage/searchPageEvents';
 import {SearchAppState} from '../../../../state/search-app-state';
-import {validatePayload} from '../../../../utils/validate-payload';
-import {
-  AnalyticsType,
-  makeAnalyticsAction,
-} from '../../../analytics/analytics-utils';
-import {
-  RangeFacetSelectionPayload,
-  rangeFacetSelectionPayloadDefinition,
-} from './range-facet-validate-payload';
+import {RangeFacetSelectionPayload} from './range-facet-validate-payload';
 
 export const getRangeFacetMetadata = (
   state: Partial<SearchAppState>,
@@ -26,22 +18,3 @@ export const getRangeFacetMetadata = (
     facetRangeStart: `${selection.start}`,
   };
 };
-
-/**
- * Logs a range facet breadcrumb event.
- * @param payload (RangeFacetSelectionPayload) Object specifying the target facet and selection.
- */
-export const logRangeFacetBreadcrumb = (payload: RangeFacetSelectionPayload) =>
-  makeAnalyticsAction(
-    'analytics/rangeFacet/breadcrumb',
-    AnalyticsType.Search,
-    (client, state) => {
-      validatePayload(
-        payload,
-        rangeFacetSelectionPayloadDefinition(payload.selection)
-      );
-      const metadata = getRangeFacetMetadata(state, payload);
-
-      return client.logBreadcrumbFacet(metadata);
-    }
-  )();
